Bring in new libhtp update to 0.4
Bring in the latest from libhtp master, it has a bunch of bug fixes. While
we are not likely to hit most of them we should still include them. The
biggest change is a new hook (response_start) which is called as soon as the
parser is ready for a response.

Add new methods to the python connection parser object for the new hook.

Add new hook to documentation.

Adjust .gitignore to be less annoying.

Bump version to 0.4.