Auto destroy transactions.
When using htpy.init() the default is to now automatically destroy
transactions after they are done being processed. If you do not want this
behavior you can create your own config, and use it. The default when
creating your own config is to have it off.

Fix a bug where we were always referencing the first transaction in the
list. This resulted in pipelined HTTP being incorrectly represented.

While here bump to 0.6.