Update libhtp to latest on master branch.
Along with bug fixes and improvements to libhtp, this brings in a number of
externally visible changes for htpy.

First is that HOOK_* constants are deprecated. They were removed from libhtp
in favor of HTP_STOP, HTP_ERROR and HTP_OK. To ensure existing htpy users
don't have to migrate right away I have shimmed support for them in with
this commit.

Second change is that the "request" and "response" hooks have been renamed
in libhtp to "request_done" and "response_done". As such I have added new
methods to the connection parser object to reflect the new names. The new
methods are "register_request_done()" and "register_response_done()". To
ensure backwards compatability I have shimmed the old names so they will
still work.

Both of these shims will be removed in a future version so please do not
rely on them and migrate any existing code you have.